extract logic into new DeleteDatetime component
